fix(audit): rename/split files to clear pre-existing audit-all errors on develop
Clears the three audit-all categories failing develop's
`test_audit.py::test_audit_all_clean`:

audit-project (PS-204 / PS-205):
- rename `tests/_state/test_preflight_creds.py` →
  `test__preflight_creds.py` (private src `_preflight_creds.py`
  requires double-underscore prefix).
- merge `tests/cli_pkg/lifecycle/test__cross_host_a2a_port.py`
  into existing siblings: dispatch null-propagation case appended
  to `test__dispatch.py::TestDispatchSshSuccessPath`; the two
  `--no-redispatch --json` a2a_port cases moved to a new
  `TestStartNoRedispatchJsonA2aPort` class in `test__start.py`.
  The cross_host_a2a_port file is then deleted (orphan: no
  matching src module).

audit-skills (SK-401):
- split `07_a2a-protocol.md` (282 lines / 14838 B) into two leaves
  both under the 200-line / 10240-B budget:
  - `07_a2a-protocol.md` keeps overview / CLI / auto-launch /
    SDK methods / handler env vars / boundary
  - `07_a2a-protocol-extension-fields.md` carries the per-agent /
    fleet / proxy `x-scitex-agent-container.*` field tables + the
    full JSON example
- register the new leaf in `SKILL.md`.

audit-mcp-tools (§6 API parity):
- add `agent.send` to `_api/agent.py.__all__` so the existing
  `agent_send` MCP tool wraps a discoverable Python API. This drops
  the orphan count from 4 to 3, below the rule's > 3 threshold.
  The remaining 3 orphans (`list_python_apis`, `quota_watch`,
  `subagent_get_state`) are meta / mis-named tools that don't
  trip the rule at count = 3.
